fix: Functionality feat: Add web support

+++ lib/ui/views/recents/recents_viewmodel.dart
@@ -1,12 +1,14 @@
+import 'package:spot/core/enums/dialog_status.dart';
 import 'package:spot/core/utils/exports.dart';
 
 class RecentsViewModel extends BaseViewModel {
   late final Box box;
 
   final NavigationService _navigationService = locator<NavigationService>();
+  final DialogService _dialogService = locator<DialogService>();
 
   Future<dynamic> initHiveBox() async {
-    box = Hive.box(HiveBoxes.searchStorageBox);
+    box = Hive.box(HiveBoxes.searchBox);
     return box;
   }
 
@@ -14,8 +16,20 @@ class RecentsViewModel extends BaseViewModel {
     HiveUtil.insertAtIndex(index);
   }
 
-  void deleteAll() {
-    HiveUtil.deleteAllData();
+  void showConfirmationDialog() async {
+    final dialogResponse = await _dialogService.showCustomDialog(
+      variant: DialogType.confirmation,
+      customData: DialogStatus.warning,
+      barrierDismissible: true,
+      title: kConfirmationTitle,
+      description: kConfirmationText,
+      mainButtonTitle: 'Yeah, totally',
+      secondaryButtonTitle: 'Nope',
+    );
+
+    if (dialogResponse?.confirmed == true) {
+      HiveUtil.deleteAllData();
+    }
   }
 
   void navigateBack() {
@@ -23,6 +37,6 @@ class RecentsViewModel extends BaseViewModel {
   }
 
   void navigateTo(route) {
-    _navigationService.navigateTo(route);
+   _navigationService.navigateTo(route);
   }
 }
